Create Readme1.md
<!-- This is the markdown template for the final project of the Building AI course, 
created by Reaktor Innovations and University of Helsinki.  -->

# Project Title

Final project for the Building AI course - Mushroom identifier

## Summary

Idea is to build a mobile app to identify mushrooms when walking in the forrest. Google Lens is similar application but not precise enough


## Background

Son is interested in mushrooms and when going to forrest it would be much easier (?) to identify correct mushroom from mobile app than going through the mushroom book with limited amount of pics and mushrooms.


## How is it used?

Solution to be used by anyone interested in mushroom. Used in forest or anywhere. 


## Data sources and AI methods
Trained data could be found from public trained algorithms and people could send pics, labeled, via app or web page. Training then would make the app more precise. 

## Challenges
To find base trained algorithm and to get people to use the app to get more pics for retraining

## What next?
Need to find algorithm as a base. 
Need top build a mobile app with AI algotihm in it

## Acknowledgments
My self and my son.